requirements.txt
# ===== requirements.txt =====
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
Pillow>=10.2.0
python-decouple==3.8